+++ packages/flutter_safe_js_platform_interface/lib/flutter_safe_js_method_channel.dart
@@ -4,6 +4,12 @@ import 'package:flutter_safe_js_platform_interface/flutter_safe_js_platform_inte
 const MethodChannel _channel = MethodChannel('plugins.flutter.io/url_launcher');
 
 class FlutterSafeJsMethodChannel extends FlutterSafeJsPlatformInterface {
+  /// @see https://stackoverflow.com/questions/50187680/how-to-call-methods-in-dart-portion-of-the-app-from-the-native-platform-using-m
+  @override
+  Future<void> expose(String javascriptFunctionName, FlutterSafeJsMethodHandler fn) async {
+    _channel.setMethodCallHandler((call) => Future.value(null));
+  }
+
   @override
   Future<void> callMethod(String javascriptFunctionName, List<dynamic> arguments) async {
     await _channel.invokeMethod<dynamic>('callMethod', {