+++ packages/flutter_safe_js/CHANGELOG.md
@@ -1,7 +1,15 @@
-# 0.0.2
+## 0.0.5
+
+ * expose handler handle only one value for compat
+
+## 0.0.3
+
+ * Add expose() handler to allow exposing a method handler by name
+ 
+## 0.0.2
 
  * add basic implementation
 
-# 0.0.1
+## 0.0.1
 
  * first version for plugin facade