+++ packages/flutter_safe_js_platform_interface/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.5
+
+ * expose handler handle only one value for compat
+
+## 0.0.2
+
+ * Add expose() handler to allow exposing a method handler by name
+
 ## 0.0.1
 
  * First root platform version